+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## 0.2.1 - [DD/MM/YYYY]
+## 0.3.0 - 24/12/2021
 
+- *Breaking*: rename methods (replace terms `Top` and `Bottom` by `Start` and `End`, because the scrollbar can be horizontal)
 - Fix `animate` parameter on differents methods (didn't work as expected)
 - Add fields (`canScroll`, `canForward`, `canBackward`)